core: extract ParamValue
This is a first step in refactoring mavlink parameters into sender and
receiver.
