+++ src/mavsdk/plugins/camera/camera_definition.cpp
@@ -307,7 +307,7 @@ CameraDefinition::parse_options(
                         return std::make_pair<>(false, options);
                     }
 
-                    MAVLinkParameters::ParamValue new_param_value;
+                    ParamValue new_param_value;
                     new_param_value.set_from_xml(
                         type_map[roption_parameter_str], roption_value_str);
                     new_parameter_range[roption_name_str] = new_param_value;
@@ -345,7 +345,7 @@ CameraDefinition::parse_range_options(
         return std::make_tuple<>(false, options, default_option);
     }
 
-    MAVLinkParameters::ParamValue min_value;
+    ParamValue min_value;
     min_value.set_from_xml(type_map[param_name], min_str);
 
     const char* max_str = param_handle->Attribute("max");
@@ -358,7 +358,7 @@ CameraDefinition::parse_range_options(
     min_option->name = "min";
     min_option->value = min_value;
 
-    MAVLinkParameters::ParamValue max_value;
+    ParamValue max_value;
     max_value.set_from_xml(type_map[param_name], max_str);
 
     auto max_option = std::make_shared<Option>();
@@ -371,7 +371,7 @@ CameraDefinition::parse_range_options(
     }
 
     if (step_str) {
-        MAVLinkParameters::ParamValue step_value;
+        ParamValue step_value;
         step_value.set_from_xml(type_map[param_name], step_str);
 
         auto step_option = std::make_shared<Option>();
@@ -392,7 +392,7 @@ CameraDefinition::parse_range_options(
         return std::make_tuple<>(false, options, default_option);
     }
 
-    MAVLinkParameters::ParamValue default_value;
+    ParamValue default_value;
     default_value.set_from_xml(type_map[param_name], default_str);
 
     default_option.name = default_str;
@@ -459,8 +459,7 @@ void CameraDefinition::assume_default_settings()
     }
 }
 
-bool CameraDefinition::get_all_settings(
-    std::unordered_map<std::string, MAVLinkParameters::ParamValue>& settings)
+bool CameraDefinition::get_all_settings(std::unordered_map<std::string, ParamValue>& settings)
 {
     std::lock_guard<std::mutex> lock(_mutex);
 
@@ -472,8 +471,7 @@ bool CameraDefinition::get_all_settings(
     return (settings.size() > 0);
 }
 
-bool CameraDefinition::get_possible_settings(
-    std::unordered_map<std::string, MAVLinkParameters::ParamValue>& settings)
+bool CameraDefinition::get_possible_settings(std::unordered_map<std::string, ParamValue>& settings)
 {
     std::lock_guard<std::mutex> lock(_mutex);
 
@@ -481,7 +479,7 @@ bool CameraDefinition::get_possible_settings(
 }
 
 bool CameraDefinition::get_possible_settings_locked(
-    std::unordered_map<std::string, MAVLinkParameters::ParamValue>& settings)
+    std::unordered_map<std::string, ParamValue>& settings)
 {
     settings.clear();
 
@@ -521,8 +519,7 @@ bool CameraDefinition::get_possible_settings_locked(
     return (settings.size() > 0);
 }
 
-bool CameraDefinition::set_setting(
-    const std::string& name, const MAVLinkParameters::ParamValue& value)
+bool CameraDefinition::set_setting(const std::string& name, const ParamValue& value)
 {
     std::lock_guard<std::mutex> lock(_mutex);
 
@@ -564,7 +561,7 @@ bool CameraDefinition::set_setting(
     return true;
 }
 
-bool CameraDefinition::get_setting(const std::string& name, MAVLinkParameters::ParamValue& value)
+bool CameraDefinition::get_setting(const std::string& name, ParamValue& value)
 {
     std::lock_guard<std::mutex> lock(_mutex);
 
@@ -582,9 +579,7 @@ bool CameraDefinition::get_setting(const std::string& name, MAVLinkParameters::P
 }
 
 bool CameraDefinition::get_option_value(
-    const std::string& param_name,
-    const std::string& option_value,
-    MAVLinkParameters::ParamValue& value)
+    const std::string& param_name, const std::string& option_value, ParamValue& value)
 {
     std::lock_guard<std::mutex> lock(_mutex);
 
@@ -603,8 +598,7 @@ bool CameraDefinition::get_option_value(
     return false;
 }
 
-bool CameraDefinition::get_all_options(
-    const std::string& name, std::vector<MAVLinkParameters::ParamValue>& values)
+bool CameraDefinition::get_all_options(const std::string& name, std::vector<ParamValue>& values)
 {
     std::lock_guard<std::mutex> lock(_mutex);
 
@@ -623,7 +617,7 @@ bool CameraDefinition::get_all_options(
 }
 
 bool CameraDefinition::get_possible_options(
-    const std::string& name, std::vector<MAVLinkParameters::ParamValue>& values)
+    const std::string& name, std::vector<ParamValue>& values)
 {
     std::lock_guard<std::mutex> lock(_mutex);
 
@@ -634,7 +628,7 @@ bool CameraDefinition::get_possible_options(
         return false;
     }
 
-    std::unordered_map<std::string, MAVLinkParameters::ParamValue> settings;
+    std::unordered_map<std::string, ParamValue> settings;
     if (!get_possible_settings_locked(settings)) {
         return false;
     }
@@ -665,7 +659,7 @@ bool CameraDefinition::get_possible_options(
     }
 
     // TODO: use set instead of vector for this
-    std::vector<MAVLinkParameters::ParamValue> allowed_ranges{};
+    std::vector<ParamValue> allowed_ranges{};
 
     // Check allowed ranges.
     for (const auto& parameter : _parameter_map) {
@@ -717,8 +711,7 @@ bool CameraDefinition::get_possible_options(
     return true;
 }
 
-void CameraDefinition::get_unknown_params(
-    std::vector<std::pair<std::string, MAVLinkParameters::ParamValue>>& params)
+void CameraDefinition::get_unknown_params(std::vector<std::pair<std::string, ParamValue>>& params)
 {
     std::lock_guard<std::mutex> lock(_mutex);
