+++ src/mavsdk/plugins/camera/camera_definition.h
@@ -27,23 +27,18 @@ public:
 
     struct Setting {
         std::string name;
-        MAVLinkParameters::ParamValue value;
+        ParamValue value;
     };
 
-    bool set_setting(const std::string& name, const MAVLinkParameters::ParamValue& value);
-    bool get_setting(const std::string& name, MAVLinkParameters::ParamValue& value);
-    bool get_all_settings(std::unordered_map<std::string, MAVLinkParameters::ParamValue>& settings);
-    bool
-    get_possible_settings(std::unordered_map<std::string, MAVLinkParameters::ParamValue>& settings);
+    bool set_setting(const std::string& name, const ParamValue& value);
+    bool get_setting(const std::string& name, ParamValue& value);
+    bool get_all_settings(std::unordered_map<std::string, ParamValue>& settings);
+    bool get_possible_settings(std::unordered_map<std::string, ParamValue>& settings);
 
     bool get_option_value(
-        const std::string& param_name,
-        const std::string& option_value,
-        MAVLinkParameters::ParamValue& value);
-    bool
-    get_all_options(const std::string& name, std::vector<MAVLinkParameters::ParamValue>& values);
-    bool get_possible_options(
-        const std::string& name, std::vector<MAVLinkParameters::ParamValue>& values);
+        const std::string& param_name, const std::string& option_value, ParamValue& value);
+    bool get_all_options(const std::string& name, std::vector<ParamValue>& values);
+    bool get_possible_options(const std::string& name, std::vector<ParamValue>& values);
 
     bool is_setting_range(const std::string& name);
 
@@ -51,8 +46,7 @@ public:
     bool get_option_str(
         const std::string& setting_name, const std::string& option_name, std::string& description);
 
-    void
-    get_unknown_params(std::vector<std::pair<std::string, MAVLinkParameters::ParamValue>>& params);
+    void get_unknown_params(std::vector<std::pair<std::string, ParamValue>>& params);
     void set_all_params_unknown();
 
     // Non-copyable
@@ -60,14 +54,13 @@ public:
     const CameraDefinition& operator=(const CameraDefinition&) = delete;
 
 private:
-    bool get_possible_settings_locked(
-        std::unordered_map<std::string, MAVLinkParameters::ParamValue>& settings);
+    bool get_possible_settings_locked(std::unordered_map<std::string, ParamValue>& settings);
 
-    typedef std::unordered_map<std::string, MAVLinkParameters::ParamValue> parameter_range_t;
+    typedef std::unordered_map<std::string, ParamValue> parameter_range_t;
 
     struct Option {
         std::string name{};
-        MAVLinkParameters::ParamValue value{};
+        ParamValue value{};
         std::vector<std::string> exclusions{};
         std::unordered_map<std::string, parameter_range_t> parameter_ranges{};
     };
@@ -77,7 +70,7 @@ private:
         bool is_control{false};
         bool is_readonly{false};
         bool is_writeonly{false};
-        MAVLinkParameters::ParamValue type{}; // for type only, doesn't hold a value
+        ParamValue type{}; // for type only, doesn't hold a value
         std::vector<std::string> updates{};
         std::vector<std::shared_ptr<Option>> options{};
         Option default_option{};
@@ -106,7 +99,7 @@ private:
     std::unordered_map<std::string, std::shared_ptr<Parameter>> _parameter_map{};
 
     struct InternalCurrentSetting {
-        MAVLinkParameters::ParamValue value{};
+        ParamValue value{};
         bool needs_updating{false};
     };
